@@ -1,6 +1,6 @@
-// Polyfill fn.bind() for PhantomJS
-/* eslint-disable no-extend-native */
-Function.prototype.bind = require('function-bind');
+import Vue from 'vue';
+
+Vue.config.productionTip = false;
 
 // require all test files (files that ends with .spec.js)
 const testsContext = require.context('./specs', true, /\.spec$/);
@@ -9,5 +9,5 @@ testsContext.keys().forEach(testsContext);
 // require all src files except main.js for coverage.
 // you can also change this to match only the subset of files that
 // you want coverage for.
-const srcContext = require.context('src', true, /^\.\/(?!main(\.js)?$)/);
+const srcContext = require.context('../../src', true, /^\.\/(?!main(\.js)?$)/);
 srcContext.keys().forEach(srcContext);
