@@ -2,5 +2,5 @@ var merge = require('webpack-merge');
 var prodEnv = require('./prod.env');
 
 module.exports = merge(prodEnv, {
-    NODE_ENV: '"development"'
+  NODE_ENV: '"development"'
 });
